+++ src/parser.ts
@@ -2250,13 +2250,15 @@ function parseCStandard(
     return isGnu ? "gnu99" : "c99";
   } else if (/(c|gnu)(11|1x|iso9899:2011)/.test(std)) {
     return isGnu ? "gnu11" : "c11";
-  } else if (/(c|gnu)(17|18|23|2x|iso9899:(2017|2018|2024))/.test(std)) {
+  } else if (/(c|gnu)(17|18|iso9899:(2017|2018))/.test(std)) {
     if (canUseGnu) {
       // cpptools supports 'c17' in same version it supports GNU std.
       return isGnu ? "gnu17" : "c17";
     } else {
       return "c11";
     }
+  } else if (/(c|gnu)(23|2x|iso9899:2024)/.test(std)) {
+    return isGnu ? "gnu23" : "c23";
   } else {
     return undefined;
   }